app/requirements.txt
annotated-types
anyio
fastapi[standard]
idna
pydantic
pydantic_core
pyodbc
python-dotenv
sniffio
SQLAlchemy
sqlmodel
starlette
tenacity
typing-inspection
typing_extensions
